refactor: :recycle: Refactor to match proposed architecture.
