+++ source/services/api/api_post_i.py
@@ -1,0 +1,12 @@
+import json
+import logging
+import os
+from abc import ABC, abstractmethod
+
+import requests
+
+
+class APIPostI(ABC):
+    @abstractmethod
+    def post_data(self, url: str) -> None:
+        pass
