+++ source/services/automation_i.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+
+class AutomationI(ABC):
+    @staticmethod
+    @abstractmethod
+    def start():
+        pass
